requirements.txt
click>=6.7
requests>=2.22.0
python-gitlab>=1.8.0
configparser
colored>=1.3.93
